saming coding style

+++ database/migrations/1675404384051_api_tokens.ts
@@ -1,3 +1,4 @@
+/* eslint-disable prettier/prettier */
 import BaseSchema from '@ioc:Adonis/Lucid/Schema'
 
 export default class extends BaseSchema {
@@ -5,11 +6,20 @@ export default class extends BaseSchema {
 
   public async up() {
     this.schema.createTable(this.tableName, (table) => {
-      table.increments('id').primary()
-      table.integer('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
-      table.string('name').notNullable()
-      table.string('type').notNullable()
-      table.string('token', 64).notNullable().unique()
+      table.increments('id')
+        .primary()
+      table.bigInteger('user_id')
+        .unsigned()
+        .references('id')
+        .inTable('users')
+        .onDelete('CASCADE')
+      table.string('name')
+        .notNullable()
+      table.string('type')
+        .notNullable()
+      table.string('token', 64)
+        .notNullable()
+        .unique()
 
       /**
        * Uses timestampz for PostgreSQL and DATETIME2 for MSSQL
